requirements.txt
anthropic>=0.40.0
apscheduler>=3.10.0
requests>=2.31.0
pyyaml>=6.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
